# contracts/http.openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "喵连 Gateway API",
    "version": "0.1.0",
    "description": "Design-stage contract weagent/1. Paths are project-owned, not Agent-native APIs. HTTPS only outside local development. JSON Schema is canonical; WebSocket frames are defined in protocol.schema.json."
  },
  "servers": [
    {
      "url": "https://gateway.example.invalid",
      "description": "Placeholder; not a deployed service"
    }
  ],
  "security": [
    {
      "UserBearer": []
    }
  ],
...
  "paths": {
...
    "/v1/bootstrap": {
      "get": {
        "operationId": "bootstrap",
        "tags": [
          "resources"
        ],
        "summary": "Consistent first-page resources and user revision",
        "responses": {
          "200": {
            "description": "Consistent first-page resources and user revision",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./protocol.schema.json#/$defs/Bootstrap"
                }
              }
            }
          },
          "default": {
            "description": "Typed error. Retry only as documented; timeout does not mean native command failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./protocol.schema.json#/$defs/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/sessions": {
      "get": {
        "operationId": "listSessions",
        "tags": [
          "resources"
        ],
        "summary": "List authorized sessions",
        "responses": {
          "200": {
            "description": "List authorized sessions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./protocol.schema.json#/$defs/SessionsPage"
                }
              }
            }
          },
          "default": {
            "description": "Typed error. Retry only as documented; timeout does not mean native command failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./protocol.schema.json#/$defs/ErrorResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/PageToken"
          },
          {
            "$ref": "#/components/parameters/Limit"
          },
          {
            "name": "nodeId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 128
            },
            "description": "nodeId"
          },
          {
            "name": "projectId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 128
            },
            "description": "projectId"
          },
          {
            "name": "agentId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 128
            },
            "description": "agentId"
          },
          {
            "name": "state",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "./protocol.schema.json#/$defs/SessionState"
            },
            "description": "state"
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 100
            },
            "description": "Escaped title search, not content/full-text search."
          },
          {
            "name": "archived",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "pinned",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ]
      },
      "post": {
        "operationId": "createSession",
        "tags": [
          "commands"
        ],
        "summary": "Create managed session and initial turn",
        "responses": {
          "202": {
            "description": "Operation receipt: confirmed means native acceptance, NOT turn completion.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./protocol.schema.json#/$defs/Operation"
                }
              }
            }
          },
          "default": {
            "description": "Typed error. Retry only as documented; timeout does not mean native command failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./protocol.schema.json#/$defs/ErrorResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "./protocol.schema.json#/$defs/CreateSession"
              }
            }
          }
        }
      }
    },
...
    "/v1/sessions/{id}/messages": {
      "post": {
        "operationId": "sendMessage",
        "tags": [
          "commands"
        ],
        "summary": "Continue the same session or explicitly queue after current turn",
        "responses": {
          "202": {
            "description": "Operation receipt: confirmed means native acceptance, NOT turn completion.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./protocol.schema.json#/$defs/Operation"
                }
              }
            }
          },
          "default": {
            "description": "Typed error. Retry only as documented; timeout does not mean native command failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./protocol.schema.json#/$defs/ErrorResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "./protocol.schema.json#/$defs/Id"
            }
          },
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "./protocol.schema.json#/$defs/SendMessage"
              }
            }
          }
        },
        "description": "Reject offline, readonly, stale-turn and capability mismatch. queue is a Node-local ordered queue, not a Gateway reconnect backlog. V1 has no steer, PTY or generic native command endpoint."
      }
    },
    "/v1/sessions/{id}/cancel": {
      "post": {
        "operationId": "cancelTurn",
        "tags": [
          "commands"
        ],
        "summary": "Cancel the specified current turn",
        "responses": {
          "202": {
            "description": "Operation receipt: confirmed means native acceptance, NOT turn completion.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./protocol.schema.json#/$defs/Operation"
                }
              }
            }
          },
          "default": {
            "description": "Typed error. Retry only as documented; timeout does not mean native command failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./protocol.schema.json#/$defs/ErrorResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "./protocol.schema.json#/$defs/Id"
            }
          },
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "./protocol.schema.json#/$defs/CancelTurn"
              }
            }
          }
        }
      }
    },
    "/v1/requests/{id}/respond": {
      "post": {
        "operationId": "respondRequest",
        "tags": [
          "commands"
        ],
        "summary": "Resolve approval or answer typed questions",
        "responses": {
          "202": {
            "description": "Operation receipt: confirmed means native acceptance, NOT turn completion.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./protocol.schema.json#/$defs/Operation"
                }
              }
            }
          },
          "default": {
            "description": "Typed error. Retry only as documented; timeout does not mean native command failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./protocol.schema.json#/$defs/ErrorResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "./protocol.schema.json#/$defs/Id"
            }
          },
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "./protocol.schema.json#/$defs/RespondRequest"
              }
            }
          }
        },
        "description": "Server also validates question/option IDs, required fields, type, uniqueness and limits against the stored request. JSON Schema validation alone is not sufficient."
      }
    },
    "/v1/operations/{id}": {
      "get": {
        "operationId": "getOperation",
        "tags": [
          "commands"
        ],
        "summary": "Reconcile a receipt without replaying its command",
        "responses": {
          "200": {
            "description": "Operation receipt: confirmed means native acceptance, NOT turn completion.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./protocol.schema.json#/$defs/Operation"
                }
              }
            }
          },
          "default": {
            "description": "Typed error. Retry only as documented; timeout does not mean native command failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./protocol.schema.json#/$defs/ErrorResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "./protocol.schema.json#/$defs/Id"
            }
          }
        ]
      }
    },
    "/v1/pairings/preview": {
      "post": {
        "operationId": "previewPairing",
        "tags": [
          "pairing"
        ],
        "summary": "Preview one-use pairing code without binding",
        "responses": {
          "200": {
            "description": "Preview one-use pairing code without binding",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./protocol.schema.json#/$defs/PairPreview"
                }
              }
            }
          },
          "default": {
            "description": "Typed error. Retry only as documented; timeout does not mean native command failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./protocol.schema.json#/$defs/ErrorResponse"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "./protocol.schema.json#/$defs/PairPreviewInput"
              }
            }
          }
        }
      }
    },
    "/v1/pairings/confirm": {
      "post": {
        "operationId": "confirmPairing",
        "tags": [
          "pairing"
        ],
        "summary": "Atomically claim unexpired preview ticket",
        "responses": {
          "200": {
            "description": "Operation receipt: confirmed means native acceptance, NOT turn completion.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./protocol.schema.json#/$defs/Operation"
                }
              }
            }
          },
          "default": {
            "description": "Typed error. Retry only as documented; timeout does not mean native command failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./protocol.schema.json#/$defs/ErrorResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "./protocol.schema.json#/$defs/PairConfirm"
              }
            }
          }
        }
      }
    },
    "/v1/nodes/{id}/revoke": {
      "post": {
        "operationId": "revokeNode",
        "tags": [
          "pairing"
        ],
        "summary": "Revoke access, identity and active Node socket",
        "responses": {
          "200": {
            "description": "Operation receipt: confirmed means native acceptance, NOT turn completion.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./protocol.schema.json#/$defs/Operation"
                }
              }
            }
          },
          "default": {
            "description": "Typed error. Retry only as documented; timeout does not mean native command failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./protocol.schema.json#/$defs/ErrorResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "./protocol.schema.json#/$defs/Id"
            }
          },
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "./protocol.schema.json#/$defs/Empty"
              }
            }
          }
        }
      }
    },
    "/v1/notifications/{id}/read": {
      "post": {
        "operationId": "markNotificationRead",
        "tags": [
          "resources"
        ],
        "summary": "Idempotently mark own notification as read",
        "responses": {
          "200": {
            "description": "Operation receipt: confirmed means native acceptance, NOT turn completion.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./protocol.schema.json#/$defs/Operation"
                }
              }
            }
          },
          "default": {
            "description": "Typed error. Retry only as documented; timeout does not mean native command failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./protocol.schema.json#/$defs/ErrorResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "./protocol.schema.json#/$defs/Id"
            }
          },
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "./protocol.schema.json#/$defs/Empty"
              }
            }
          }
        }
      }
    },
    "/v1/changes": {
      "get": {
        "operationId": "listChanges",
        "tags": [
          "recovery"
        ],
        "summary": "Page user metadata changes to a fixed barrier",
        "responses": {
          "200": {
            "description": "Page user metadata changes to a fixed barrier",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./protocol.schema.json#/$defs/ChangesPage"
                }
              }
            }
          },
          "default": {
            "description": "Typed error. Retry only as documented; timeout does not mean native command failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./protocol.schema.json#/$defs/ErrorResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "after",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "maximum": 9007199254740991
            },
            "description": "after"
          },
          {
            "name": "until",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "maximum": 9007199254740991
            },
            "description": "until"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200,
              "default": 100
            },
            "description": "limit"
          }
        ]
      }
    },
...
    "/v1/sessions/{id}/checkpoint": {
      "get": {
        "operationId": "getCheckpoint",
        "tags": [
          "recovery"
        ],
        "summary": "Pin an immutable full-timeline checkpoint",
        "responses": {
          "200": {
            "description": "Pin an immutable full-timeline checkpoint",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./protocol.schema.json#/$defs/Checkpoint"
                }
              }
            }
          },
          "default": {
            "description": "Typed error. Retry only as documented; timeout does not mean native command failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./protocol.schema.json#/$defs/ErrorResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "./protocol.schema.json#/$defs/Id"
            }
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "UserBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Opaque user token from wx.login exchange; never put in URL."
      },
      "NodeBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Short-lived token obtained by Ed25519 proof, only valid for Node control WSS."
      },
      "EnrollmentPoll": {
        "type": "http",
        "scheme": "bearer",
        "description": "One-enrollment polling secret; cannot access client resources."
      }
    },
    "parameters": {
      "IdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": true,
        "description": "Client operation ID, 16–128 ASCII letters/digits/_/-. Scoped to user; same key and same canonical request return original receipt, changed body is 409.",
        "schema": {
          "type": "string",
          "pattern": "^[A-Za-z0-9_-]{16,128}$"
        }
      },
      "PageToken": {
        "name": "pageToken",
        "in": "query",
        "schema": {
          "type": "string",
          "maxLength": 1024
        },
        "description": "Opaque signed owner/filter-bound keyset cursor; not an event cursor."
      },
      "Limit": {
        "name": "limit",
        "in": "query",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 100,
          "default": 50
        }
      }
    }
  },
...
}